fixed code warning

+++ frameworks/common/src/avsession_pixel_map_adapter.cpp
@@ -45,7 +45,7 @@ void AVSessionPixelMapAdapter::CleanAVSessionPixelMap(std::shared_ptr<AVSessionP
 void ConvertImageInfoFromInner(std::vector<uint8_t>& innerImgBuffer, Media::ImageInfo& imageInfo)
 {
     int32_t innerImgSize = static_cast<int>(innerImgBuffer.size());
-    CHECK_AND_RETURN_LOG(innerImgSize > (IMAGE_BYTE_SIZE + sizeof(Media::ImageInfo)),
+    CHECK_AND_RETURN_LOG(innerImgSize > static_cast<int32_t>(IMAGE_BYTE_SIZE + sizeof(Media::ImageInfo)),
         "innerPixelMap has no imageInfo");
     size_t offset = IMAGE_BYTE_SIZE;
     errno_t ret = memcpy_s(&imageInfo.size, sizeof(Media::Size),
